Feature/pwm lib

+++ src/include/io.h
@@ -8,7 +8,7 @@
 #include <stdint.h>
 
 typedef struct OUT {
-    uint8_t brake;
+    uint16_t brake;
     uint16_t vout;
     uint8_t leds;
 } output_t;
